==> notebooks/1_clean_component_failure_Analysis.ipynb <==
# Group Assignment - Surviving the Analysis - Gr-54

## EDA

1. Separate data into each component failure data type and cencored data
2. Check for impact of env components - Carpets, Pets, usage Patterns
3. Check for obs which had failure but none of the 3 components failed (Censored)
4. More than one part failed when sent for repair. Check how to handle it
5. Check hours of use and impact on the sensors
   1. about 30% of the dataset had pets
   2. 30% of those who had pets had failures
   3. for the ones with battery failures, 50% had pets
   4. 30% had pets for ir failures
   5. however, over 87% of impact failures happened to those with pets. This makes sense cuz pets keep on bumping into shit.
   6. Battery failure increases (30%) when there are pets. Probably cuz they shed.
6. Check if having pets increases the time the robo runs. Could explain the battery part.
7. Check impact of carpet and pets on failures

### Assumptions:

1. Survival of each component is independent of the other
2. if one part fails, the others that did not fail go into censored data??
3. Only one part fails at a time. the 2 rows where 2 parts failed have to be handled.

## Modelling

1. Do 4 models - Exponential, weibull, kaplan-mayer, and Cox
   1. Start with a very simple model - consider dec 31 as the censoring time for everything
   2. Next iteration, consider censoring time for component x as failure time of component y
2. Do all forumals, assumptions, all dists n all
3. Handling the changing censoring time
   1. If we use models that consider constant censoring time
      1. Filter out all other failures
   2. Using models that can handle changing survival time,
      1. separate them out
      2. Set the failure time of component 1 as censoring time of the other components

## Inference

1. Check the info provided for confidence interval

#### Members:<br>

Boglarka ACS<br>
Chavisha Arora<br>
Sander Bergman<br>
Aniket Mishra<br>


## Random thoughts to consider:

.


import pandas as pd
import re
from varname import nameof

import warnings

warnings.filterwarnings("ignore")

def sanitize(input_str: str) -> str:
    return re.sub(r"[^A-Za-z0-9]+", "_", input_str)

df = pd.read_csv("../data/DirtSlurper3000.csv", skiprows=13)
df.head(10)

df.describe(include="all").T

df.info(memory_usage="deep")

df.columns = [sanitize(col).lower() for col in df.columns]
df.head(2)

df.loc[df["failure_date"] == "---", "failure_date"] = "1Jan1970"
df["registration_date"] = pd.to_datetime(df["registration_date"])
df["failure_date"] = pd.to_datetime(df["failure_date"])

df = df.sort_values(["registration_date"]).reset_index(drop=True)

df.head(1)

for col in df.columns:
    all_unique = sorted(df[col].unique())
    len_unique = len(all_unique)
    if len_unique < 10:
        print(col)
        print(all_unique)
        print()

replace_values_with_int = {"---": 0, "Damage": 1, "OK": 2, "NO": 0, "YES": 1}

df[["sent_for_repair", "battery_status", "impact_status", "ir_status"]] = df[
    ["sent_for_repair", "battery_status", "impact_status", "ir_status"]
].replace(replace_values_with_int)

df["had_failure"] = 0
df["time_to_failure"] = None
# df["time_to_failure_battery_status"] = None
# df["time_to_failure_impact_status"] = None
# df["time_to_failure_ir_status"] = None
df["type_of_failure"] = "Censored"

df.loc[df["failure_date"] != "1970-01-01", "had_failure"] = 1
mask = df["had_failure"] == 1
df.loc[mask, "time_to_failure"] = (df["failure_date"] - df["registration_date"]).dt.days
df["time_to_failure"].fillna(-1, inplace=True)

df.loc[((df["had_failure"] == 1) & (df["battery_status"] == 1)), "type_of_failure"] = (
    "Battery"
)
df.loc[((df["had_failure"] == 1) & (df["impact_status"] == 1)), "type_of_failure"] = (
    "Impact"
)
df.loc[((df["had_failure"] == 1) & (df["ir_status"] == 1)), "type_of_failure"] = "Ir"

df.loc[
    ((df["had_failure"] == 1) & (df["ir_status"] == 1) & (df["battery_status"] == 1)),
    "type_of_failure",
] = "Battery_and_Ir"
df.loc[
    ((df["had_failure"] == 1) & (df["ir_status"] == 1) & (df["impact_status"] == 1)),
    "type_of_failure",
] = "Ir_and_Impact"
df.loc[
    (
        (df["had_failure"] == 1)
        & (df["battery_status"] == 1)
        & (df["impact_status"] == 1)
    ),
    "type_of_failure",
] = "Battery_and_Impact"
df.loc[
    ((df["had_failure"] == 1) & (df["type_of_failure"] == "Censored")),
    "type_of_failure",
] = "Unknown_Component"

df.loc[df["type_of_failure"] == "Unknown_Component"].shape

## Sanity Checks - EDA


### Checking if we missed to fill any failures


df.loc[df["type_of_failure"].isna()]

df.loc[df["type_of_failure"].str.contains("and")]

### Checking for Had a failure but not sent for repair


df.loc[((df["sent_for_repair"] == 0) & (df["had_failure"] == 1))]

### Multiple components failed at the same time

## Context:

if battery is below 80% (Not a failure) but is sent to repair cuz another component failed.<br>
Manufacturer guarantees 2400 hours at 80%, so this can be considered a failure. tho people will not notice this.

i.e.<br> If battery is marked as failed along with other components, it means the health of the battry was below 80% when it went to get fixed due to the other sensor issue. We might wanna account for that. So whenever we see battery_and_othersensor, we consider both as failed and failure time as the failure time for both components.

#### Reason: Contractual - Mentioned in documentation.


df.loc[
    ((df["battery_status"] == 1) & (df["impact_status"] == 1))
    | ((df["battery_status"] == 1) & (df["ir_status"] == 1))
    | ((df["ir_status"] == 1) & (df["impact_status"] == 1))
]

### Failure on one device but other 2 devices are censored (0) and not No failure (2)


df.loc[
    (
        (df["battery_status"] == 1)
        & ((df["impact_status"] == 0) | (df["ir_status"] == 0))
    )
    | (
        (df["impact_status"] == 1)
        & ((df["battery_status"] == 0) | (df["ir_status"] == 0))
    )
    | (
        (df["ir_status"] == 1)
        & ((df["impact_status"] == 0) | (df["battery_status"] == 0))
    )
]

### Had failure but all 3 components were marked as not failed (Unknown component failure)


df.loc[
    (df["had_failure"] == 1)
    & (df["battery_status"] == 2)
    & (df["impact_status"] == 2)
    & (df["ir_status"] == 2)
].shape

### Context:

We may consider the above cases where a part other than the other 3 failed.<br>
We could use the failure time as censoring time for the 3 components.


### Had failure but all 3 components were censored (0)


df.loc[
    (df["had_failure"] == 1)
    & (df["battery_status"] == 0)
    & (df["impact_status"] == 0)
    & (df["ir_status"] == 0)
].shape

## Separating the dataset into different failures (No longer needed)


df_battery_failure = df.loc[df["battery_status"] == 1]
df_impact_failure = df.loc[df["impact_status"] == 1]
df_ir_failure = df.loc[df["ir_status"] == 1]

### Checking info on all component failures

#### Battery Failure


df_battery_failure.describe(include="all").T

df_battery_failure.info()

#### Impact Failure


df_impact_failure.describe(include="all").T

df_impact_failure.info()

#### IR Failure


df_ir_failure.describe(include="all").T

df_ir_failure.info()

#### Failure percentages


total_failure_percentage = (df.loc[df["had_failure"] == 1].shape[0] / df.shape[0]) * 100
battery_failure_percentage = (df_battery_failure.shape[0] / df.shape[0]) * 100
impact_failure_percentage = (df_impact_failure.shape[0] / df.shape[0]) * 100
ir_failure_percentage = (df_ir_failure.shape[0] / df.shape[0]) * 100

total_rows = df.shape[0]
print(f"{nameof(total_rows)}: {total_rows}")
var_dict = {
    "Total Failure": [
        {nameof(total_failure_percentage): total_failure_percentage},
        df.loc[df["had_failure"] == 1].shape[0],
        df["registration_date"].min(),
        df["registration_date"].max(),
    ],
    nameof(df_battery_failure): [
        {nameof(battery_failure_percentage): battery_failure_percentage},
        df_battery_failure.shape[0],
        df_battery_failure["registration_date"].min(),
        df_battery_failure["registration_date"].max(),
    ],
    nameof(df_impact_failure): [
        {nameof(impact_failure_percentage): impact_failure_percentage},
        df_impact_failure.shape[0],
        df_impact_failure["registration_date"].min(),
        df_impact_failure["registration_date"].max(),
    ],
    nameof(df_ir_failure): [
        {nameof(ir_failure_percentage): ir_failure_percentage},
        df_ir_failure.shape[0],
        df_ir_failure["registration_date"].min(),
        df_ir_failure["registration_date"].max(),
    ],
}

for key, percent in var_dict.items():
    print(
        f"\n{key}\n{percent[0]}, \nRows: {percent[1]}, \nMin Time:{percent[2]}, \nMax Time:{percent[3]}\n"
    )

## Getting Censored Data - No Failures at all


df_censored = df.loc[df["had_failure"] == 0]

df_censored

# Impact of env factors on failures


result = (
    df_battery_failure.groupby("carpet_score")
    .agg(
        {
            "had_failure": "count",
        }
    )
    .reset_index()
)

result

result = (
    df_impact_failure.groupby("carpet_score")
    .agg(
        {
            "had_failure": "count",
        }
    )
    .reset_index()
)

result

result = (
    df_ir_failure.groupby("carpet_score").agg({"had_failure": "count"}).reset_index()
)

result

df["pets"].unique()

df_battery_failure.loc[df_battery_failure["pets"] == 1]

df.loc[df["pets"] == 1].shape

df.loc[(df["pets"] == 1) & (df["had_failure"] == 1)].shape

df_battery_failure.shape

df.shape

df_battery_failure.shape

print(df_ir_failure.loc[df_ir_failure["pets"] == 1].shape)
print(df_ir_failure.shape)

print(df_impact_failure.loc[df_impact_failure["pets"] == 1].shape)
print(df_impact_failure.shape)

114 / 130

df.loc[(df["had_failure"] == 1) & (df["pets"] == 1)]["total_usage_time"].sum()

print(
    df_battery_failure.loc[
        (df_battery_failure["had_failure"] == 1) & (df_battery_failure["pets"] == 1)
    ]["total_usage_time"].sum()
)
print(
    df_battery_failure.loc[
        (df_battery_failure["had_failure"] == 1) & (df_battery_failure["pets"] == 0)
    ]["total_usage_time"].sum()
)
print(
    df_battery_failure.loc[(df_battery_failure["had_failure"] == 1)][
        "total_usage_time"
    ].sum()
)

print(
    df_ir_failure.loc[
        (df_ir_failure["had_failure"] == 1) & (df_ir_failure["pets"] == 1)
    ]["total_usage_time"].sum()
)
print(
    df_ir_failure.loc[
        (df_ir_failure["had_failure"] == 1) & (df_ir_failure["pets"] == 0)
    ]["total_usage_time"].sum()
)
print(df_ir_failure.loc[(df_ir_failure["had_failure"] == 1)]["total_usage_time"].sum())

print(
    df_impact_failure.loc[
        (df_impact_failure["had_failure"] == 1) & (df_impact_failure["pets"] == 1)
    ]["total_usage_time"].sum()
)
print(
    df_impact_failure.loc[
        (df_impact_failure["had_failure"] == 1) & (df_impact_failure["pets"] == 0)
    ]["total_usage_time"].sum()
)
print(
    df_impact_failure.loc[(df_impact_failure["had_failure"] == 1)][
        "total_usage_time"
    ].sum()
)

print(
    df_impact_failure.loc[
        (df_impact_failure["had_failure"] == 1) & (df_impact_failure["pets"] == 1)
    ]["total_usage_time"].count()
)
print(
    df_impact_failure.loc[
        (df_impact_failure["had_failure"] == 1) & (df_impact_failure["pets"] == 0)
    ]["total_usage_time"].count()
)
print(
    df_impact_failure.loc[(df_impact_failure["had_failure"] == 1)][
        "total_usage_time"
    ].count()
)

result = (
    df_impact_failure.loc[df_impact_failure["pets"] == 0]
    .groupby("carpet_score")
    .agg(
        {
            "had_failure": "count",
        }
    )
    .reset_index()
)

result

df.to_csv("../output/data_cleaned.csv", index=False)
df_battery_failure.to_csv("../output/df_battery_failure.csv", index=False)
df_impact_failure.to_csv("../output/df_impact_failure.csv", index=False)
df_ir_failure.to_csv("../output/df_ir_failure.csv", index=False)

df.loc[df["battery_status"] == 2, "battery_status"] = 0
df.loc[df["impact_status"] == 2, "impact_status"] = 0
df.loc[df["ir_status"] == 2, "ir_status"] = 0
df.to_csv("../output/cleaned_df_with_status_0_or_1.csv", index = False)

df.head(10)

df.columns

df.loc[df["type_of_failure"] == "Unknown_Component"].shape
