==> notebooks/4_checks.ipynb <==
import pandas as pd

df = pd.read_csv(
    "/Users/paniket/TU_Eindhoven/2_Study/Q1_2AMS11_survival_Statistics_for_Data_Scientists/2_Assignments/group/Group_54_2AMS11_Assignment/output/data_cleaned.csv"
)

df_bat = pd.read_csv(
    "/Users/paniket/TU_Eindhoven/2_Study/Q1_2AMS11_survival_Statistics_for_Data_Scientists/2_Assignments/group/Group_54_2AMS11_Assignment/output/df_battery_failure.csv"
)
df_ir = pd.read_csv(
    "/Users/paniket/TU_Eindhoven/2_Study/Q1_2AMS11_survival_Statistics_for_Data_Scientists/2_Assignments/group/Group_54_2AMS11_Assignment/output/df_ir_failure.csv"
)
df_impact = pd.read_csv(
    "/Users/paniket/TU_Eindhoven/2_Study/Q1_2AMS11_survival_Statistics_for_Data_Scientists/2_Assignments/group/Group_54_2AMS11_Assignment/output/df_impact_failure.csv"
)

df_bat.drop(columns=["battery_status", "impact_status", "ir_status"], inplace=True)
df_ir.drop(columns=["battery_status", "impact_status", "ir_status"], inplace=True)
df_impact.drop(columns=["battery_status", "impact_status", "ir_status"], inplace=True)

df_ir["time_to_failure"].mean()

df_impact.loc[df_impact["pets"] == 1]["time_to_failure"].mean()

df_impact.loc[df_impact["pets"] == 0]["time_to_failure"].mean()

df.columns

df_num = df[
    [
        "type_of_failure",
        "pets",
        "carpet_score",
        "had_failure",
        "time_to_failure",
        "total_usage_time",
    ]
]
key_cols = ["type_of_failure", "pets", "carpet_score", "had_failure"]
for col in df_num.columns:
    if col == "type_of_failure":
        continue
    df_num[col] = pd.to_numeric(df_num[col])

grouped_df = pd.DataFrame(
    columns=["type_of_failure", "pets", "carpet_score", "had_failure"]
)

agg_df_list = []

for col in df_num.columns:
    if col in key_cols:
        continue
    df_groupby = df_num.groupby(key_cols, group_keys=True)
    for agg in ["mean", "count", "max", "min", "std"]:
        colname = f"{col}_{agg}"
        print(colname)
        if agg == "mean":
            df_grouped = df_groupby.mean().reset_index()
        elif agg == "count":
            df_grouped = df_groupby.count().reset_index()
        elif agg == "max":
            df_grouped = df_groupby.max().reset_index()
        elif agg == "min":
            df_grouped = df_groupby.min().reset_index()
        elif agg == "std":
            df_grouped = df_groupby.std().reset_index()
        if df_grouped.empty:
            continue
        df_grouped.rename(columns={col: colname}, inplace=True)
        cols = key_cols + [colname]
        df_grouped = df_grouped[cols]
        # agg_df_list.append(df_grouped)
        grouped_df = grouped_df.merge(df_grouped, on=key_cols, how="outer")


# agg_df = pd.concat(agg_df_list)
# grouped_df = grouped_df.merge(agg_df, on=key_cols, how="outer")

grouped_df.head(10)

df_num = df[
    [
        "type_of_failure",
        "pets",
        "carpet_score",
        "had_failure",
        "time_to_failure",
        "total_usage_time",
    ]
]
key_cols = ["type_of_failure", "pets", "had_failure"]
for col in df_num.columns:
    if col == "type_of_failure":
        continue
    df_num[col] = pd.to_numeric(df_num[col])
    
    grouped_df = pd.DataFrame(
    columns=["type_of_failure", "pets", "had_failure"]
)

agg_df_list = []

for col in df_num.columns:
    if col in key_cols:
        continue
    df_groupby = df_num.groupby(key_cols, group_keys=True)
    for agg in ["mean", "count", "max", "min", "std"]:
        colname = f"{col}_{agg}"
        print(colname)
        if agg == "mean":
            df_grouped = df_groupby.mean().reset_index()
        elif agg == "count":
            df_grouped = df_groupby.count().reset_index()
        elif agg == "max":
            df_grouped = df_groupby.max().reset_index()
        elif agg == "min":
            df_grouped = df_groupby.min().reset_index()
        elif agg == "std":
            df_grouped = df_groupby.std().reset_index()
        if df_grouped.empty:
            continue
        df_grouped.rename(columns={col: colname}, inplace=True)
        cols = key_cols + [colname]
        df_grouped = df_grouped[cols]
        # agg_df_list.append(df_grouped)
        grouped_df = grouped_df.merge(df_grouped, on=key_cols, how="outer")


# agg_df = pd.concat(agg_df_list)
# grouped_df = grouped_df.merge(agg_df, on=key_cols, how="outer")

grouped_df.columns

grouped_df[['type_of_failure', 'pets', 'had_failure', 'time_to_failure_mean', 'time_to_failure_count',
       'time_to_failure_max', 'time_to_failure_min', 'time_to_failure_std',
       ]]

grouped_df[['type_of_failure', 'pets', 'had_failure', 'total_usage_time_mean', 'total_usage_time_count',
       'total_usage_time_max', 'total_usage_time_min', 'total_usage_time_std'   
       ]]

grouped_df.to_csv(
    "../output/aggregated_information_for_each_failure_grouped.csv", index=False
)

# grouped_df["had_failure_mean"].unique()

# Ignore below, prototyping


df_grouped = (
    df_num.groupby(["type_of_failure", "pets", "carpet_score"], group_keys=True)
    .count()
    .reset_index()
)
df_grouped2 = (
    df_num.groupby(["type_of_failure", "pets", "carpet_score"], group_keys=True)
    .mean()
    .reset_index()
)
df_grouped3 = (
    df_num.groupby(["type_of_failure", "pets", "carpet_score"], group_keys=True)
    .min()
    .reset_index()
)
df_grouped4 = (
    df_num.groupby(["type_of_failure", "pets", "carpet_score"], group_keys=True)
    .max()
    .reset_index()
)
df_grouped5 = (
    df_num.groupby(["type_of_failure", "pets", "carpet_score"], group_keys=True)
    .std()
    .reset_index()
)

df_grouped

df_grouped2

df_grouped3

df_grouped4

df_grouped5

df

df["registration_date"] = pd.to_datetime(df["registration_date"])
df["failure_date"] = pd.to_datetime(df["failure_date"])

df

df.describe().T

df.loc[df["had_failure"] == 1].describe().T

df.loc[df["had_failure"] == 1].groupby("type_of_failure").agg("mean").reset_index(drop=False).describe()

df.loc[(df["had_failure"] == 1) & (df["battery_status"] == 1)].describe().T
